requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.1.0
langchain-tavily>=0.1.0
python-dotenv>=1.0.0
sendgrid>=6.11.0
streamlit>=1.35.0